requirements.txt
Pillow==8.3.1
numpy==1.20.3
matplotlib==3.3.4
torch==1.9.0
torchvision==0.10.0
streamlit==0.85.1
streamlit-drawable-canvas==0.8.0